Refactored code into modules
