+++ api/endpoints.js
@@ -1,0 +1,9 @@
+const BASE_URL = "https://pokeapi.co/api/v2/";
+
+export const API_ENDPOINTS = {
+  POKEMON: `${BASE_URL}pokemon/`,
+  ABILITY: `${BASE_URL}ability/`,
+  SPECIES: `${BASE_URL}pokemon-species/`,
+  TYPE: `${BASE_URL}type/`,
+  MOVE: `${BASE_URL}move/`
+};
